requirements.txt
pdf2image==1.14.0
img2pdf==0.4.0
PyPDF2==1.26.0
google-cloud-dlp==3.0.1
google-cloud-storage==1.36.2
functions-framework==2.1.2
